Add GitHub workflow for Conan WASM package creation
A new GitHub workflow file has been added to automate the process of creating WASM packages with Conan. This workflow includes steps for setting up Conan and dependencies, caching packages, and uploading the generated package to the appropriate Conan repository. Furthermore, secrets are used to securely handle access and authentication to Conan repositories.

Contribute to NP-202