Pass secret along in order to be able to sign the uninstaller.
The uninstaller is generated during the packaging, so we can't really sign it here without a workaround. Fortunately, the nsis stuff we use has a built-in command that sort-of handles this -- but we do need to make an argument to pass the secret along we use for the signing.

part of CURA-12129