Stop uploading built binary packages
Building the packages is actually useful to check that they do build, but in most situations it is very likely that we won't ever reuse them. Thus we decided not to upload them every time, in order to save bandwidth and storage space on the artifactory. On the other hand, we will keep uploading binary packages built during installers built, because these are ponctual actions and more likely to be reused. For projects that don't have installers, it should be done when merging a feature branch to main.