temp workaround: Version of the msvc redist shipped with Qt doesn't work with Arcus.
We found out that the only reason our install is even working reliably is because it takes the redistributable files from the Qt/bin folder from PyQt. After an upgrade however, we found out that there's a 'technically not a breaking change' in a slightly newer version of the entire MSVC system that broke Arcus if not built with the right version. We'll make a ticket later to see if it even can be cleaned up better -- but the only viable plan so far would be to scan all the directories, build a C(++) program with each of them that prints out the _MSVC_STL_VERSION for each of them, and for all that match that, take the highest. -- All other roads investigated led to dissapointment.

CURA-12421